requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# OpenAI Integration
openai==1.55.3

# Google Sheets Integration
gspread==6.1.4
google-auth==2.36.0

# HTTP Client
httpx==0.28.1
aiohttp==3.9.1

# WebSocket Support
websockets==12.0

# File Upload Support
python-multipart==0.0.20

# Environment Variables
python-dotenv==1.0.1

# Audio Processing (for format conversion)
pydub==0.25.1

# Testing (optional for development)
pytest==8.3.4
pytest-asyncio==0.24.0
